Huge refactor, I am sorry about the huge commit. Basically I try to remove the need of cast and group all similar methods together to avoid repetition. Change the xml format, noticed an increase on execution time (due to testng executing test by order found) but I think is better organized now
