+++ src/test/java/test/clients/posts/BasePostClient.java
@@ -1,8 +1,13 @@
 package test.clients.posts;
 
+import pojo.PostRequestBody;
 import test.clients.BaseClient;
 
 public abstract class BasePostClient extends BaseClient {
+
+    protected PostRequestBody postObject;
+    protected int postId;
+    protected int userId;
 
     public BasePostClient(String baseUri) {
         super(baseUri);
@@ -25,5 +30,18 @@ public abstract class BasePostClient extends BaseClient {
     public int getUserId(){
         return getApiResponseAsJsonObject().getInt("userId");
     }
+
+    public void setPostObject(PostRequestBody postObject){
+        this.postObject = postObject;
+    }
+
+    public void setPostId(int postId){
+        this.postId = postId;
+    }
+
+    public void setUserId(int userId){
+        this.userId = userId;
+    }
+
 
 }
