+++ src/test/java/test/clients/posts/PostByUserClient.java
@@ -6,17 +6,11 @@ import test.utils.JsonUtilities;
 
 import java.util.List;
 
-public class PostByUserClient extends BaseClient {
-
-    private int userId;
+public class PostByUserClient extends BasePostClient {
 
     public PostByUserClient(String baseUri){
         super(baseUri);
         apiClient = new PostByUserApi(baseUri);
-    }
-
-    public void setUserId(int userId){
-        this.userId = userId;
     }
 
     @Override
