+++ src/test/java/test/clients/posts/UpdatePostClient.java
@@ -2,24 +2,12 @@ package test.clients.posts;
 
 import myTaxy.apiModels.posts.UpdateAPostApi;
 import pojo.PostRequestBody;
-import test.clients.BaseClient;
 
-public class UpdatePostClient extends BaseClient {
-
-    private PostRequestBody postObject;
-    private int postId;
+public class UpdatePostClient extends BasePostClient {
 
     public UpdatePostClient(String baseUri){
         super(baseUri);
         apiClient = new UpdateAPostApi(baseUri);
-    }
-
-    public void setPostObject(PostRequestBody postObject){
-        this.postObject = postObject;
-    }
-
-    public void setPostId(int postId){
-        this.postId = postId;
     }
 
     @Override
